:truck: Rename abc to dataclasses
