+++ bot/__init__.py
@@ -1,6 +1,6 @@
 """AutomaTime telegram bot"""
 
-from bot.abc import Session, CompleteSession
+from bot.dataclasses import Session, CompleteSession
 
 # Global constants
 START_CODE = "#START"
